Create blog page server side rendering issue fixed
