+++ app/loading.jsx
@@ -1,13 +1,9 @@
 "use client";
 
-import { PulseLoader } from "react-spinners";
+import Loader from "@/components/shared/Loader";
 
 const loading = () => {
-  return (
-    <div className="w-screen h-screen overflow-hidden flex justify-center items-center">
-      <PulseLoader color="#ffc306" size={15} speedMultiplier={0.6} />
-    </div>
-  );
+  return <Loader />;
 };
 
 export default loading;
